Create sub signer and move it into metadata namespace.
