+++ osu.Game.Rulesets.Karaoke/Beatmaps/Metadatas/Types/ISinger.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Rulesets.Karaoke.Beatmaps.Metadatas.Types
+{
+    public interface ISinger
+    {
+    }
+}
