+++ osu.Game.Rulesets.Karaoke/Statistics/BeatmapMetadataGraph.cs
@@ -16,6 +16,7 @@ using osu.Game.Rulesets.Karaoke.Skinning.Components;
 using osuTK;
 using osuTK.Graphics;
 using System.Linq;
+using osu.Game.Rulesets.Karaoke.Beatmaps.Metadatas;
 
 namespace osu.Game.Rulesets.Karaoke.Statistics
 {
